Add files via upload
As an aspiring SOC analyst, I was both excited and a bit nervous to dive into TryHackMe’s SOC Level 1 training — my first structured course for the role. The training offered a broad and practical introduction to the tools, frameworks, and techniques used in real-world security operations. I expected to work with tools like SIEMs, Wireshark, and Snort, but I was pleasantly surprised to explore many others I hadn’t encountered before, such as Autopsy, Zeek, and OSQuery. Getting hands-on with such a wide range of technologies made the experience both challenging and rewarding.